+++ R/tm_g_distribution.R
@@ -109,21 +109,14 @@ tm_g_distribution <- function(label = "Distribution Module",
                               pre_output = NULL,
                               post_output = NULL) {
   logger::log_info("Initializing tm_g_distribution")
-  if (!requireNamespace("ggpmisc", quietly = TRUE)) {
-    stop("Cannot load ggpmisc - please install the package or restart your session.")
+
+  extra_packages <- c("ggpmisc", "ggpp", "goftest", "MASS", "broom")
+  missing_packages <- Filter(function(x) !requireNamespace(x, quietly = TRUE), extra_packages)
+  if (length(missing_packages) > 0L) {
+    stop(sprintf("Cannot load package(s): %s.\nInstall or restart your session.",
+                 paste(missing_packages, sep = ", ")))
   }
-  if (!requireNamespace("ggpp", quietly = TRUE)) {
-    stop("Cannot load ggpp - please install the package or restart your session.")
-  }
-  if (!requireNamespace("goftest", quietly = TRUE)) {
-    stop("Cannot load goftest - please install the package or restart your session.")
-  }
-  if (!requireNamespace("MASS", quietly = TRUE)) {
-    stop("Cannot load MASS - please install the package or restart your session.")
-  }
-  if (!requireNamespace("broom", quietly = TRUE)) {
-    stop("Cannot load broom - please install the package or restart your session.")
-  }
+
   if (inherits(dist_var, "data_extract_spec")) dist_var <- list(dist_var)
   if (inherits(strata_var, "data_extract_spec")) strata_var <- list(strata_var)
   if (inherits(group_var, "data_extract_spec")) group_var <- list(group_var)
@@ -172,13 +165,6 @@ ui_distribution <- function(id, ...) {
   ns <- NS(id)
   is_single_dataset_value <- teal.transform::is_single_dataset(args$dist_var, args$strata_var, args$group_var)
 
-  scales_condition_base <- paste0(
-    "input['",
-    extract_input(ns("group_i"), args$group_var[[1]]$dataname, filter = TRUE),
-    "']"
-  )
-  scales_condition <- paste0(scales_condition_base, " && ", paste0(scales_condition_base, ".length != 0"))
-
   teal.widgets::standard_layout(
     output = tagList(
       tabsetPanel(
@@ -211,17 +197,7 @@ ui_distribution <- function(id, ...) {
             data_extract_spec = args$group_var,
             is_single_dataset = is_single_dataset_value
           ),
-          conditionalPanel(
-            condition = scales_condition,
-            shinyWidgets::prettyRadioButtons(
-              ns("scales_type"),
-              label = "Scales:",
-              choices = c("Fixed", "Free"),
-              selected = "Fixed",
-              bigger = FALSE,
-              inline = TRUE
-            )
-          )
+          uiOutput(ns("scales_types_ui"))
         )
       },
       if (!is.null(args$strata_var)) {
@@ -293,14 +269,14 @@ ui_distribution <- function(id, ...) {
           "Tests:",
           choices = c(
             "Shapiro-Wilk",
-            "t-test (two-samples, not paired)",
-            "one-way ANOVA",
-            "Fligner-Killeen",
-            "F-test",
+            if (!is.null(args$strata_var)) "t-test (two-samples, not paired)",
+            if (!is.null(args$strata_var)) "one-way ANOVA",
+            if (!is.null(args$strata_var)) "Fligner-Killeen",
+            if (!is.null(args$strata_var)) "F-test",
             "Kolmogorov-Smirnov (one-sample)",
             "Anderson-Darling (one-sample)",
             "Cramer-von Mises (one-sample)",
-            "Kolmogorov-Smirnov (two-samples)"
+            if (!is.null(args$strata_var)) "Kolmogorov-Smirnov (two-samples)"
           ),
           selected = NULL
         )
@@ -343,9 +319,101 @@ srv_distribution <- function(id,
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
   checkmate::assert_class(data, "tdata")
   moduleServer(id, function(input, output, session) {
-    data_extract <- list(dist_i = dist_var, strata_i = strata_var, group_i = group_var)
 
-    selector_list <- teal.transform::data_extract_multiple_srv(data_extract, data)
+    rule_req <- function(value) {
+      if (isTRUE(input$dist_tests %in% c("Fligner-Killeen",
+                                         "t-test (two-samples, not paired)",
+                                         "F-test",
+                                         "Kolmogorov-Smirnov (two-samples)",
+                                         "one-way ANOVA"))) {
+        if (!shinyvalidate::input_provided(value))
+          "Please select stratify variable."
+      }
+    }
+    rule_dupl <- function(...) {
+      if (identical(input$dist_tests, "Fligner-Killeen")) {
+        strata <- selector_list()$strata_i()$select
+        group <- selector_list()$group_i()$select
+        if (isTRUE(strata == group))
+          "Please select different variables for strata and group."
+      }
+    }
+
+    selector_list <- teal.transform::data_extract_multiple_srv(
+      data_extract = list(
+        dist_i = dist_var,
+        strata_i = strata_var,
+        group_i = group_var
+      ),
+      data,
+      select_validation_rule = list(
+        dist_i = shinyvalidate::sv_required("Please select a variable")
+      ),
+      filter_validation_rule = list(
+        strata_i = shinyvalidate::compose_rules(
+          rule_req,
+          rule_dupl
+        ),
+        group_i = rule_dupl
+      )
+    )
+
+    iv_r <- reactive({
+      iv <- shinyvalidate::InputValidator$new()
+      teal.transform::compose_and_enable_validators(iv, selector_list, validator_names = "dist_i")
+    })
+
+    iv_r_dist <- reactive({
+      iv <- shinyvalidate::InputValidator$new()
+      teal.transform::compose_and_enable_validators(
+        iv, selector_list, validator_names = c("strata_i", "group_i"))
+    })
+    rule_dist_1 <- function(value) {
+      if (!is.null(input$t_dist)) {
+        switch(
+          input$t_dist,
+          "normal" = if (!shinyvalidate::input_provided(value)) "mean is required",
+          "lognormal" = if (!shinyvalidate::input_provided(value)) "meanlog is required",
+          "gamma" = {
+            if (!shinyvalidate::input_provided(value)) "shape is required" else
+              if (value <= 0) "shape must be positive"
+          },
+          "unif" = NULL)
+      }
+    }
+    rule_dist_2 <- function(value) {
+      if (!is.null(input$t_dist)) {
+        switch(
+          input$t_dist,
+          "normal" = {
+            if (!shinyvalidate::input_provided(value)) "sd is required" else
+              if (value < 0) "sd must be non-negative"
+          },
+          "lognormal" = {
+            if (!shinyvalidate::input_provided(value)) "sdlog is required" else
+              if (value < 0) "sdlog must be non-negative"
+          },
+          "gamma" = {
+            if (!shinyvalidate::input_provided(value)) "rate is required" else
+              if (value <= 0) "rate must be positive"
+          },
+          "unif" = NULL)
+      }
+    }
+    rule_dist <- function(value) {
+      if (isTRUE(input$tabs == "QQplot" ||
+                 input$dist_tests %in% c("Kolmogorov-Smirnov (one-sample)",
+                                         "Anderson-Darling (one-sample)",
+                                         "Cramer-von Mises (one-sample)"))) {
+        if (!shinyvalidate::input_provided(value))
+          "Please select the theoretical distribution."
+      }
+    }
+    iv_dist <- shinyvalidate::InputValidator$new()
+    iv_dist$add_rule("t_dist", rule_dist)
+    iv_dist$add_rule("dist_param1", rule_dist_1)
+    iv_dist$add_rule("dist_param2", rule_dist_2)
+    iv_dist$enable()
 
     anl_merged_input <- teal.transform::merge_expression_srv(
       selector_list = selector_list,
@@ -364,11 +432,24 @@ srv_distribution <- function(id,
       anl_q_r = anl_merged_q
     )
 
+    output$scales_types_ui <- renderUI({
+      if ("group_i" %in% names(selector_list()) && length(selector_list()$group_i()$filters[[1]]$selected) > 0) {
+        shinyWidgets::prettyRadioButtons(
+          session$ns("scales_type"),
+          label = "Scales:",
+          choices = c("Fixed", "Free"),
+          selected = "Fixed",
+          bigger = FALSE,
+          inline = TRUE
+        )
+      }
+    })
+
     observeEvent(
       eventExpr = list(
         input$t_dist,
         input$params_reset,
-        input[[extract_input("dist_i", dist_var[[1]]$dataname)]]
+        selector_list()$dist_i()$select
       ),
       handlerExpr = {
         if (length(input$t_dist) != 0) {
@@ -402,6 +483,8 @@ srv_distribution <- function(id,
     )
 
     merge_vars <- reactive({
+      teal::validate_inputs(iv_r())
+
       dist_var <- as.vector(merged$anl_input_r()$columns_source$dist_i)
       s_var <- as.vector(merged$anl_input_r()$columns_source$strata_i)
       g_var <- as.vector(merged$anl_input_r()$columns_source$group_i)
@@ -423,9 +506,6 @@ srv_distribution <- function(id,
     # common qenv
     common_q <- reactive({
       # Create a private stack for this function only.
-      validate({
-        need(length(merged$anl_input_r()$columns_source$dist_i) > 0, "Please select a variable")
-      })
 
       ANL <- merged$anl_q_r()[["ANL"]] # nolint
       dist_var <- merge_vars()$dist_var
@@ -480,17 +560,13 @@ srv_distribution <- function(id,
       teal::validate_has_data(ANL, 1, complete = TRUE)
 
       if (length(t_dist) != 0) {
-        map_distr_nams <- data.frame(
-          distr = c("normal", "lognormal", "gamma", "unif"),
-          namparam = I(list(
-            c("mean", "sd"),
-            c("meanlog", "sdlog"),
-            c("shape", "rate"),
-            c("min", "max")
-          )),
-          stringsAsFactors = FALSE
+        map_distr_nams <- list(
+          normal = c("mean", "sd"),
+          lognormal = c("meanlog", "sdlog"),
+          gamma = c("shape", "rate"),
+          unif = c("min", "max")
         )
-        params_names_raw <- map_distr_nams$namparam[match(t_dist, map_distr_nams$distr)][[1]]
+        params_names_raw <- map_distr_nams[[t_dist]]
 
         qenv <- teal.code::eval_code(
           qenv,
@@ -569,9 +645,6 @@ srv_distribution <- function(id,
         is.null(input$ggtheme)
       },
       valueExpr = {
-        ANL <- common_q()[["ANL"]] # nolint
-        summary_table <- common_q()[["summary_table"]] # nolint
-
         dist_var <- merge_vars()$dist_var
         s_var <- merge_vars()$s_var
         g_var <- merge_vars()$g_var
@@ -583,52 +656,50 @@ srv_distribution <- function(id,
         dist_param2 <- input$dist_param2
 
         scales_type <- input$scales_type
+
         ndensity <- 512
         main_type_var <- input$main_type
         bins_var <- input$bins
         add_dens_var <- input$add_dens
         ggtheme <- input$ggtheme
 
-        validate(need(ggtheme, "Please select a theme."))
+        teal::validate_inputs(iv_dist)
 
         qenv <- common_q()
 
-        m_type <- if (main_type_var == "Density") "..density.." else "..count.."
-        m_type2 <- if (main_type_var == "Density") {
-          "..density.."
-        } else {
-          paste(diff(range(ANL[[dist_var]], na.rm = TRUE)) / bins_var, "* ..count..")
-        }
+        m_type <- if (main_type_var == "Density") "density" else "count"
 
         plot_call <- if (length(s_var) == 0 && length(g_var) == 0) {
           substitute(
             expr = ggplot(ANL, aes(dist_var_name)) +
-              geom_histogram(position = "identity", aes_string(y = m_type), bins = bins_var, alpha = 0.3),
+              geom_histogram(
+                position = "identity", aes(y = after_stat(m_type)), bins = bins_var, alpha = 0.3),
             env = list(
-              m_type = m_type,
-              bins_var = bins_var,
-              dist_var_name = as.name(dist_var)
+              m_type = as.name(m_type), bins_var = bins_var, dist_var_name = as.name(dist_var)
             )
           )
         } else if (length(s_var) != 0 && length(g_var) == 0) {
           substitute(
             expr = ggplot(ANL, aes(dist_var_name, col = s_var_name)) +
-              geom_histogram(position = "identity", aes_string(y = m_type, fill = s_var), bins = bins_var, alpha = 0.3),
+              geom_histogram(
+                position = "identity", aes(y = after_stat(m_type), fill = s_var), bins = bins_var, alpha = 0.3),
             env = list(
-              m_type = m_type,
+              m_type = as.name(m_type),
               bins_var = bins_var,
               dist_var_name = dist_var_name,
-              s_var = s_var,
+              s_var = as.name(s_var),
               s_var_name = s_var_name
             )
           )
         } else if (length(s_var) == 0 && length(g_var) != 0) {
+          req(scales_type)
           substitute(
             expr = ggplot(ANL[ANL[[g_var]] != "NA", ], aes(dist_var_name)) +
-              geom_histogram(position = "identity", aes_string(y = m_type), bins = bins_var, alpha = 0.3) +
+              geom_histogram(
+                position = "identity", aes(y = after_stat(m_type)), bins = bins_var, alpha = 0.3) +
               facet_wrap(~g_var_name, ncol = 1, scales = scales_raw),
             env = list(
-              m_type = m_type,
+              m_type = as.name(m_type),
               bins_var = bins_var,
               dist_var_name = dist_var_name,
               g_var = g_var,
@@ -637,19 +708,20 @@ srv_distribution <- function(id,
             )
           )
         } else {
+          req(scales_type)
           substitute(
             expr = ggplot(ANL[ANL[[g_var]] != "NA", ], aes(dist_var_name, col = s_var_name)) +
               geom_histogram(
                 position = "identity",
-                aes_string(y = m_type, fill = s_var), bins = bins_var, alpha = 0.3
+                aes(y = after_stat(m_type), fill = s_var), bins = bins_var, alpha = 0.3
               ) +
               facet_wrap(~g_var_name, ncol = 1, scales = scales_raw),
             env = list(
-              m_type = m_type,
+              m_type = as.name(m_type),
               bins_var = bins_var,
               dist_var_name = dist_var_name,
               g_var = g_var,
-              s_var = s_var,
+              s_var = as.name(s_var),
               g_var_name = g_var_name,
               s_var_name = s_var_name,
               scales_raw = tolower(scales_type)
@@ -661,7 +733,7 @@ srv_distribution <- function(id,
           plot_call <- substitute(
             expr = plot_call +
               stat_density(
-                aes_string(y = m_type2),
+                aes(y = after_stat(const * m_type2)),
                 geom = "line",
                 position = "identity",
                 alpha = 0.5,
@@ -670,13 +742,15 @@ srv_distribution <- function(id,
               ),
             env = list(
               plot_call = plot_call,
-              m_type2 = m_type2,
+              const = if (main_type_var == "Density") 1 else
+                diff(range(qenv[["ANL"]][[dist_var]], na.rm = TRUE)) / bins_var,
+              m_type2 = if (main_type_var == "Density") as.name("density") else as.name("count"),
               ndensity = ndensity
             )
           )
         }
 
-        if (length(t_dist) != 0 && m_type == "..density.." && length(g_var) == 0 && length(s_var) == 0) {
+        if (length(t_dist) != 0 && main_type_var == "Density" && length(g_var) == 0 && length(s_var) == 0) {
           qenv <- teal.code::eval_code(
             qenv,
             substitute(
@@ -697,8 +771,8 @@ srv_distribution <- function(id,
           )
         }
 
-        if (length(s_var) == 0 && length(g_var) == 0 && m_type == "..density.." &&
-          length(t_dist) != 0 && m_type == "..density..") {
+        if (length(s_var) == 0 && length(g_var) == 0 && main_type_var == "Density" &&
+            length(t_dist) != 0 && main_type_var == "Density") {
           map_dist <- stats::setNames(
             c("dnorm", "dlnorm", "dgamma", "dunif"),
             c("normal", "lognormal", "gamma", "unif")
@@ -755,9 +829,6 @@ srv_distribution <- function(id,
         is.null(input$ggtheme)
       },
       valueExpr = {
-        ANL <- common_q()[["ANL"]] # nolint
-        summary_table <- common_q()[["summary_table"]]
-
         dist_var <- merge_vars()$dist_var
         s_var <- merge_vars()$s_var
         g_var <- merge_vars()$g_var
@@ -771,26 +842,19 @@ srv_distribution <- function(id,
         scales_type <- input$scales_type
         ggtheme <- input$ggtheme
 
-        validate(need(ggtheme, "Please select a theme."))
-        validate(need(t_dist, "Please select the theoretical distribution."))
-        validate_dist_parameters(t_dist, dist_param1, dist_param2)
+        teal::validate_inputs(iv_r_dist(), iv_dist)
 
         qenv <- common_q()
 
         plot_call <- if (length(s_var) == 0 && length(g_var) == 0) {
           substitute(
             expr = ggplot(ANL, aes_string(sample = dist_var)),
-            env = list(
-              dist_var = dist_var
-            )
+            env = list(dist_var = dist_var)
           )
         } else if (length(s_var) != 0 && length(g_var) == 0) {
           substitute(
             expr = ggplot(ANL, aes_string(sample = dist_var, color = s_var)),
-            env = list(
-              dist_var = dist_var,
-              s_var = s_var
-            )
+            env = list(dist_var = dist_var, s_var = s_var)
           )
         } else if (length(s_var) == 0 && length(g_var) != 0) {
           substitute(
@@ -860,10 +924,7 @@ srv_distribution <- function(id,
           plot_call <- substitute(
             expr = plot_call +
               stat_qq_line(distribution = mapped_dist, dparams = params),
-            env = list(
-              plot_call = plot_call,
-              mapped_dist = as.name(unname(map_dist[t_dist]))
-            )
+            env = list(plot_call = plot_call, mapped_dist = as.name(unname(map_dist[t_dist])))
           )
         }
 
@@ -919,6 +980,8 @@ srv_distribution <- function(id,
 
         validate(need(dist_tests, "Please select a test"))
 
+        teal::validate_inputs(iv_dist)
+
         if (length(s_var) > 0 || length(g_var) > 0) {
           counts <- ANL %>%
             dplyr::group_by_at(dplyr::vars(dplyr::any_of(c(s_var, g_var)))) %>%
@@ -927,21 +990,12 @@ srv_distribution <- function(id,
           validate(need(all(counts$n > 5), "Please select strata*group with at least 5 observation each."))
         }
 
+
         if (dist_tests %in% c(
-          "Kolmogorov-Smirnov (one-sample)",
-          "Anderson-Darling (one-sample)",
-          "Cramer-von Mises (one-sample)"
-        )) {
-          validate(need(t_dist, "Please select the theoretical distribution."))
-        } else if (dist_tests == "Fligner-Killeen") {
-          validate(need(s_var, "Please select stratify variable."))
-          validate(need(!identical(s_var, g_var), "Please select different variables for strata and group."))
-        } else if (dist_tests %in% c(
           "t-test (two-samples, not paired)",
           "F-test",
           "Kolmogorov-Smirnov (two-samples)"
         )) {
-          validate(need(s_var, "Please select stratify variable."))
           if (length(g_var) == 0 && length(s_var) > 0) {
             validate(need(
               length(unique(ANL[[s_var]])) == 2,
@@ -950,14 +1004,12 @@ srv_distribution <- function(id,
           }
           if (length(g_var) > 0 && length(s_var) > 0) {
             validate(need(
-              all(stats::na.omit(as.vector(tapply(
-                ANL[[s_var]], list(ANL[[g_var]]), function(x) length(unique(x))
-              ) == 2))),
+              all(stats::na.omit(as.vector(
+                tapply(ANL[[s_var]], list(ANL[[g_var]]), function(x) length(unique(x))) == 2))
+              ),
               "Please select stratify variable with 2 levels, per each group."
             ))
           }
-        } else if (dist_tests == "one-way ANOVA") {
-          validate(need(s_var, "Please select stratify variable."))
         }
 
         map_dist <- stats::setNames(
@@ -1011,15 +1063,15 @@ srv_distribution <- function(id,
         )
 
         tests_base <- switch(dist_tests,
-          "Kolmogorov-Smirnov (one-sample)" = sks_args,
-          "Shapiro-Wilk" = ssw_args,
-          "Fligner-Killeen" = mfil_args,
-          "one-way ANOVA" = manov_args,
-          "t-test (two-samples, not paired)" = mt_args,
-          "F-test" = mv_args,
-          "Kolmogorov-Smirnov (two-samples)" = mks_args,
-          "Anderson-Darling (one-sample)" = sad_args,
-          "Cramer-von Mises (one-sample)" = scvm_args
+                             "Kolmogorov-Smirnov (one-sample)" = sks_args,
+                             "Shapiro-Wilk" = ssw_args,
+                             "Fligner-Killeen" = mfil_args,
+                             "one-way ANOVA" = manov_args,
+                             "t-test (two-samples, not paired)" = mt_args,
+                             "F-test" = mv_args,
+                             "Kolmogorov-Smirnov (two-samples)" = mks_args,
+                             "Anderson-Darling (one-sample)" = sad_args,
+                             "Cramer-von Mises (one-sample)" = scvm_args
         )
 
         env <- list(
@@ -1093,21 +1145,24 @@ srv_distribution <- function(id,
       qenv_final
     })
 
-
     dist_r <- reactive(dist_q()[["g"]])
 
     qq_r <- reactive(qq_q()[["g"]])
 
-    tests_r <- reactive(test_q()[["test_stats"]])
-
     output$summary_table <- DT::renderDataTable(
-      expr = common_q()[["summary_table"]],
+      expr = if (iv_r()$is_valid()) common_q()[["summary_table"]] else NULL,
       options = list(
         autoWidth = TRUE,
         columnDefs = list(list(width = "200px", targets = "_all"))
       ),
       rownames = FALSE
     )
+
+    tests_r <- reactive({
+      req(iv_r()$is_valid())
+      teal::validate_inputs(iv_r_dist())
+      test_q()[["test_stats"]]
+    })
 
     pws1 <- teal.widgets::plot_with_settings_srv(
       id = "hist_plot",
@@ -1177,34 +1232,4 @@ srv_distribution <- function(id,
     }
     ###
   })
-}
-
-#' @description
-#' Validates the parameters of the given theoretical distribution.
-#'
-#' @note Returns a Shiny validation error if the parameters don't meet
-#' the assumptions of the theoretical distribution.
-#'
-#' @param dist_type (`character(1)`) the family of a distribution
-#' @param dist_param1 (`numeric(1)`) the first parameter of the distribution
-#' @param dist_param2 (`numeric(1)`) the second parameter of the distribution
-#' @return NULL
-#' @noRd
-validate_dist_parameters <- function(dist_type, dist_param1, dist_param2) {
-  switch(dist_type,
-    "normal" = {
-      validate(need(dist_param2 >= 0, "Variance of the normal distribution needs to be nonnegative"))
-    },
-    "lognormal" = {
-      validate(need(dist_param2 >= 0, "Sigma parameter of the log-normal distribution needs to be nonnegative"))
-    },
-    "gamma" = {
-      validate(need(
-        dist_param1 > 0 && dist_param2 > 0,
-        "k and theta parameters of the gamma distribution need to be positive"
-      ))
-    },
-    "unif" = NULL
-  )
-  NULL
 }