+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.modules.general 0.2.15.9012
+# teal.modules.general 0.2.15.9013
 
 ### Breaking changes
 
@@ -11,6 +11,7 @@
 * Improved UI labels and plot panel title in `tm_g_association`.
 * Added inputs `tm_variable_browser` module for text size and plot theme.
 * Forced ggplot theme to be always selected in all modules.
+* Updated encodings input checks to use `shinyvalidate::InputValidator` instead of `shiny::validate` for better UI experience.   
 
 ### Bug fixes
 * Fixed a bug in `tm_g_scatterplot` when selected x and y facets were the same.
@@ -20,6 +21,7 @@
 ### Miscellaneous
 
 * Examples now use `scda.2022` rather than `scda.2021`.
+* Replaced deprecated `ggplot2` functions `..count..`, `..density..` and `..prop..`.
 
 # teal.modules.general 0.2.15